LayoutTests/http/tests/navigation/page-cache-expired-entry.html

<!-- webkit-test-runner [ UsesBackForwardCache=true ] -->
<!DOCTYPE html>
<html>
<body>
<script src="/resources/js-test-pre.js"></script>
<script>
description("Tests that we're not crashing when navigating back to a back/forward entry after WebInspector has disabled caches.");
window.jsTestIsAsync = true;

window.addEventListener("pageshow", function(event) {
    debug("pageshow - " + (event.persisted ? "" : "not ") + "from cache");

    if (sessionStorage.didStartPageCacheExpiredEntryTest) {
      shouldBeFalse("event.persisted");
      setTimeout(finishJSTest, 0);
    }
}, false);

window.addEventListener("pagehide", function(event) {
    debug("pagehide - " + (event.persisted ? "" : "not ") + "entering cache");
    if (!event.persisted) {
        testFailed("Page did not enter the page cache.");
        finishJSTest();
    }
}, false);

window.addEventListener('load', function() {
    if (sessionStorage.didStartPageCacheExpiredEntryTest)
      return;

    // This needs to happen in a setTimeout because a navigation inside the onload handler would
    // not create a history entry.
    setTimeout(function() {
      sessionStorage.didStartPageCacheExpiredEntryTest = true;
      // Force a back navigation back to this page.
      location.href = "resources/page-cache-helper-inspector-disables-caches.html";
    }, 0);
}, false);

</script>
<script src="/resources/js-test-post.js"></script>
</body>
</html>
